requirements.txt
# Workshop participant requirements
python-dotenv
wandb-workspaces
wandb
torch
torchvision
timm
scikit-learn
numpy
Pillow
tqdm
nbformat

# Setup script only (admin)
datasets
